+++ src/model_io/codecs/src/JointElement.cpp
@@ -9,7 +9,6 @@
 #include <iDynTree/XMLAttribute.h>
 #include <iDynTree/XMLParser.h>
 
-#include <iDynTree/Model/IJoint.h>
 #include <iDynTree/Model/FixedJoint.h>
 #include <iDynTree/Model/PrismaticJoint.h>
 #include <iDynTree/Model/RevoluteJoint.h>
@@ -123,6 +122,35 @@ namespace iDynTree {
                 return true;
             });
             return std::shared_ptr<XMLElement>(element);
+
+        } else if (name == "dynamics") {
+            m_dynamic_params = std::make_shared<JointDynamicsParams>();
+            m_dynamic_params->jointDynamicsType = URDFJointDynamics;
+            m_dynamic_params->damping = .0;
+            m_dynamic_params->staticFriction = .0;
+
+            // TODO: check how the defaults/required works
+            XMLElement* element = new XMLElement(getParserState(), name);
+            element->setAttributeCallback(
+                [this](const std::unordered_map<std::string, std::shared_ptr<XMLAttribute>>& attributes) {
+                    auto found = attributes.find("damping");
+                    if (found != attributes.end()) {
+                        double value = 0;
+                        if (stringToDoubleWithClassicLocale(found->second->value(), value)) {
+                            m_dynamic_params->damping = value;
+                        }
+                    }
+                    found = attributes.find("friction");
+                    if (found != attributes.end()) {
+                        double value = 0;
+                        if (stringToDoubleWithClassicLocale(found->second->value(), value)) {
+                            m_dynamic_params->staticFriction = value;
+                        }
+                    }
+                    return true;
+                    }
+                );
+            return std::shared_ptr<XMLElement>(element);
         }
         return std::make_shared<XMLElement>(getParserState(), name);
     }
@@ -166,7 +194,13 @@ namespace iDynTree {
                 std::string errStr = "Joint " + m_jointName + " misses the limit tag.";
                 reportWarning("JointElement", "", errStr.c_str());
             }
-            
+
+            if(m_dynamic_params) {
+                info.joint->setJointDynamicsType(URDFJointDynamics);
+                info.joint->setDamping(0, m_dynamic_params->damping);
+                info.joint->setStaticFriction(0, m_dynamic_params->staticFriction);
+            }
+
             if (!map->insert(std::unordered_map<std::string, JointElement::JointInfo>::value_type(m_jointName, info)).second) {
                 std::string errStr = "Duplicate joint " + m_jointName + " found.";
                 reportError("JointElement", "", errStr.c_str());