+++ src/model_io/codecs/src/ModelExporter.cpp
@@ -14,7 +14,8 @@ namespace iDynTree
 ModelExporterOptions::ModelExporterOptions()
 : baseLink(""),
     exportFirstBaseLinkAdditionalFrameAsFakeURDFBase(true),
-    robotExportedName("iDynTreeURDFModelExportModelName") {}
+    robotExportedName("iDynTreeURDFModelExportModelName"),
+    xmlBlobs{} {}
 
 
 class ModelExporter::Pimpl {